Fix budget bugs

+++ src/main/java/com/spendwise/servlet/BudgetServlet.java
@@ -17,28 +17,30 @@ public class BudgetServlet extends HttpServlet {
 
         HttpSession session = req.getSession(false);
         if (session == null || session.getAttribute("userId") == null) {
-            res.sendRedirect("login.jsp"); return;
+            res.sendRedirect("login.jsp");
+            return;
         }
 
-        int userId = (int) session.getAttribute("userId");
+        int userId    = (int) session.getAttribute("userId");
         LocalDate now = LocalDate.now();
-        int month = now.getMonthValue();
-        int year  = now.getYear();
+        int month     = now.getMonthValue();
+        int year      = now.getYear();
 
-        BudgetDAO dao = new BudgetDAO();
-        BigDecimal budget = dao.getBudget(userId, month, year);
-        BigDecimal spent  = dao.getTotalSpent(userId, month, year);
+        BudgetDAO dao    = new BudgetDAO();
+        BigDecimal budget    = dao.getBudget(userId, month, year);
+        BigDecimal spent     = dao.getTotalSpent(userId, month, year);
         BigDecimal remaining = budget.subtract(spent);
+
         double percent = budget.doubleValue() > 0
-            ? Math.min(100, (spent.doubleValue() / budget.doubleValue()) * 100)
-            : 0;
+            ? Math.min(100.0, (spent.doubleValue() / budget.doubleValue()) * 100.0)
+            : 0.0;
 
         req.setAttribute("budget",    budget);
         req.setAttribute("spent",     spent);
         req.setAttribute("remaining", remaining);
-        req.setAttribute("percent",   Math.round(percent));
-        req.setAttribute("month",     month);
-        req.setAttribute("year",      year);
+        req.setAttribute("percent",   (int) Math.round(percent));
+        req.setAttribute("month",     (Integer) month);
+        req.setAttribute("year",      (Integer) year);
 
         req.getRequestDispatcher("budget.jsp").forward(req, res);
     }
@@ -49,13 +51,14 @@ public class BudgetServlet extends HttpServlet {
 
         HttpSession session = req.getSession(false);
         if (session == null || session.getAttribute("userId") == null) {
-            res.sendRedirect("login.jsp"); return;
+            res.sendRedirect("login.jsp");
+            return;
         }
 
-        int userId = (int) session.getAttribute("userId");
+        int userId       = (int) session.getAttribute("userId");
         BigDecimal amount = new BigDecimal(req.getParameter("budgetAmount"));
-        int month = Integer.parseInt(req.getParameter("month"));
-        int year  = Integer.parseInt(req.getParameter("year"));
+        int month        = Integer.parseInt(req.getParameter("month"));
+        int year         = Integer.parseInt(req.getParameter("year"));
 
         new BudgetDAO().setBudget(userId, month, year, amount);
         res.sendRedirect("budget?msg=saved");
