+++ src/main/java/com/spendwise/servlet/ReportServlet.java
@@ -16,23 +16,28 @@ public class ReportServlet extends HttpServlet {
 
         HttpSession session = req.getSession(false);
         if (session == null || session.getAttribute("userId") == null) {
-            res.sendRedirect("login.jsp"); return;
+            res.sendRedirect("login.jsp");
+            return;
         }
 
-        int userId = (int) session.getAttribute("userId");
+        int userId    = (int) session.getAttribute("userId");
         LocalDate now = LocalDate.now();
 
         int month = req.getParameter("month") != null
-            ? Integer.parseInt(req.getParameter("month")) : now.getMonthValue();
-        int year  = req.getParameter("year") != null
-            ? Integer.parseInt(req.getParameter("year"))  : now.getYear();
+            ? Integer.parseInt(req.getParameter("month"))
+            : now.getMonthValue();
+
+        int year = req.getParameter("year") != null
+            ? Integer.parseInt(req.getParameter("year"))
+            : now.getYear();
 
         ReportDAO dao = new ReportDAO();
+
         req.setAttribute("categoryTotals",     dao.getCategoryTotals(userId, month, year));
         req.setAttribute("bankCategoryTotals", dao.getBankCategoryTotals(userId, month, year));
         req.setAttribute("monthlyTrend",       dao.getMonthlyTrend(userId));
-        req.setAttribute("month", month);
-        req.setAttribute("year",  year);
+        req.setAttribute("month",              (Integer) month);
+        req.setAttribute("year",               (Integer) year);
 
         req.getRequestDispatcher("reports.jsp").forward(req, res);
     }
